Refactor and rename classes for better readability
