+++ src/main/java/quizhub/command/CommandEdit.java
@@ -1,7 +1,7 @@
 package quizhub.command;
 
 import quizhub.storage.Storage;
-import quizhub.quizlist.QuizList;
+import quizhub.questionlist.QuizList;
 import quizhub.ui.Ui;
 
 public class CommandEdit extends Command {
