+++ src/main/java/quizhub/question/ShortAnsQn.java
@@ -1,14 +1,14 @@
-package quizhub.quiz;
+package quizhub.question;
 /**
  * Represents todo tasks in the current task list
  */
-public class ShortAnsQn extends Quiz {
+public class ShortAnsQn extends Question {
     /**
      * Creates a new task of type todo.
      *
      * @param description Task description.
      */
     public ShortAnsQn(String description){
-        super(description, TaskType.TODO);
+        super(description, TaskType.SHORTANSWER);
     }
 }
